Updates to WordCloudCore and WordCloudTextProcess. Also I added the functionality which allows to run the whole application with "docker compose up -d".
WordCloudTextProcess is basically finished (Microservice currently doesn't have the possibility to let user choose the words that should be omitted).  Also added a small functionality which checks if the word has possible spelling error and it saves into the data base.

WordCloudCore is also basically finished. It allows frontend application to upload text file which will be processed by the WordCloudTextProcess service. It also allows frontend application to get the results which has been stored to the database by the WordCloudTextProcess. Currently, it doesn't have the possibility to  let user choose the words that should be omitted and the possibility to see what is the status of the text processing.